Add tests and refactor to allow testing

+++ src/tools.rs
@@ -1,4 +1,4 @@
-pub fn get_input(day: u8) -> String {
+pub fn get_input(day: usize) -> String {
     let filename = format!("./assets/day{}.txt", day);
     std::fs::read_to_string(filename).unwrap()
 }
